Address review feedback: rename internal package to clientauth to avoid naming conflict
