+++ extension/googleclientauthextension/internal/clientauth/testdata/fake_creds_no_project.json
