--- extension/googleclientauthextension/internal/googleclientauthextension/doc.go
@@ -1,5 +1,0 @@
-// Copyright The OpenTelemetry Authors
-// SPDX-License-Identifier: Apache-2.0
-
-// Package googleclientauthextension provides authentication using Google client libraries.
-package googleclientauthextension // import "github.com/open-telemetry/opentelemetry-collector-contrib/extension/googleclientauthextension/internal/googleclientauthextension"
